#328 refactor : 관심 종목 추가, 이름 수정, drawer 반응형 추가
